requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pinecone-client
langchain-google-genai
langchain-pinecone
google-generativeai
mysql-connector-python
beautifulsoup4
selenium
requests
streamlit
pydantic
